refactor: 토큰 생성을 위한 비즈니스 로직 흐름 정돈

+++ src/main/java/com/lof/auth/domain/InvalidRefreshToken.java
@@ -22,9 +22,9 @@ public class InvalidRefreshToken {
     private long memberId;
 
     @TimeToLive
-    private Long expirationInSeconds;
+    private Integer expirationInSeconds;
 
-    public InvalidRefreshToken(ValidRefreshToken validRefreshToken, Long expirationInSeconds) {
+    public InvalidRefreshToken(ValidRefreshToken validRefreshToken, int expirationInSeconds) {
         this(validRefreshToken.getRefreshToken(), validRefreshToken.getId(), expirationInSeconds);
     }
 }
